Combined GPS polling. Put serial radio stuff in new serial.Comm include file. Standardized most include names.
